requirements.txt
Flask
Pillow
pandas
matplotlib
numpy 
